Added permission requirements to appointments
You must now have a role in the system in order to create or delete
appointments.

+++ delete-appointment.php
@@ -1,6 +1,18 @@
 <?php
+session_start();
 if ($_SERVER["REQUEST_METHOD"] == "POST") {
   require_once "connection.php";
+
+  $userID = $_SESSION["userid"];
+  $sql = "SELECT Role AS role FROM users WHERE UserID = '$userID';";
+  $result = $conn->query($sql);
+  $userRole = $result->fetch_assoc()['role'];
+  $roles = array("Administrator", "Doctor", "Nurse");
+  if (!in_array($userRole, $roles)) {
+    Header("Location: appointments.php?delete-appointment-msg=permission");
+    exit();
+  }
+
   $name = $_POST['Pname'];
   $date = $_POST['Date'];
   $sql = "DELETE FROM appointments WHERE Pname = '$name' && Date = '$date';";
